# [Frontend] Reduce frequency of client cancellation checking