The api server logic currently checks for client http connection disconnection for each request between every generated token. We can reduce the cost associated with this by limiting the checks to at most once per second.

With llama-3.2-1B, A100 tp=1, `--num-scheduler-steps=8`:
| stream | multistep-stream | req/sec before this PR | req/sec after this PR |
| -- | -- | -- | -- |
| false | n/a | 95 | 95 |
| true | false | 92 | 94 |
| true | true | 42 | 67 |
```
python benchmarks/benchmark_serving.py \
    --backend vllm \
    --model meta-llama/Llama-3.2-1B \
    --dataset-name sharegpt \
    --dataset-path ShareGPT_V3_unfiltered_cleaned_split.json \
    --port 9999 \
    --save-result \
    --result-dir results \
    --result-filename test.json \
    --num-prompts 2000 \
    --request-rate 200 \
    --seed 42
```